all: Separate band package definitions
